Update check to use async Mutex
Adding capability to have async tests so that we can make outgoing http calls to endpoints.
This requires async health check.
We provide a non-async health check called FFIProbe to handle the FFI health probes.
Limitation is that async probes can only be generated as part of the HaMS library as they are not
FFI safe.
Through the implementation this required passing the MutexGuard across the async boundary.
The std::Mutex does not support Sync/Send so cannot be shared over async boundary.
Tokio Async Mutex allows passing of mutexguard over the async boundary. The previous mutex
was converted to a Tokio:: Mutex.
As a result this required the use of blocking_lock for insert/remove BUT these do no work
within the async threads. This should not be an issue for the FFI interfaces as they are
blocking and outside the async threds. BUT they do block the tests. So create a wrapper (scoped to tests)
that allows an async thread to also update the checks using the blocking call.

NOTE; If we wanted to use async directly this would not be necessary but we want to test via teh blocking
code so that the test is representive and has coverage for the FFI exposed functions.